% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_causaleffects.R
\name{get_causaleffects}
\alias{get_causaleffects}
\title{Estimation of the (average) causal effects under intervention on the exposures}
\usage{
get_causaleffects(
  output,
  ord = NULL,
  BMA = TRUE,
  CI = 0.95,
  progress = FALSE,
  plot = FALSE
)
}
\arguments{
\item{output}{Output produced by \code{MrDAG} algorithm}

\item{ord}{Indices of the traits in \code{data} that define the outcomes and the exposures. They specify the order of appearance
of the traits when printing the (average) causal effect. If \code{ord = NULL}, the order is the same as \code{MrDAGcheck} in
\code{MrDAG} ouput}

\item{BMA}{If \code{TRUE}, Bayesian Model Averaging of the estimated (average) causal effect across all explored DAGs in the
visited Completed Partially DAG (CPDAG) (\insertCite{Chickering2002;textual}{MrDAG}) or Essential Graph (EG)
(\insertCite{Andersson1997;textual}{MrDAG}) is performed}

\item{CI}{Level (\code{0.95} default) of the credible interval (CI) of the (average) causal effect. It is calculated based on
suitable quantiles of the estimated (average) causal effect across all explored DAGs}

\item{progress}{Logical value set as \code{FALSE} by default to print on the screen the progress of the causal effects estimation}

\item{plot}{Logical value (default \code{FALSE}). If \code{TRUE}, \code{get_causaleffects} generates the output to be used in a
plot}
}
\value{
The value returned is a list object \code{list(causalEffects, causalEffects_LL, causalEffects_UL, group, ord, BMA, CI)}
\itemize{
\item{\code{causalEffects}}{ Estimate of the (average) causal effects under intervention on the exposures based on the DAGs
explored by \code{MrDAG} algorithm. Feedback loop causal effect of the same outcome is not allowed (NA) }
\item{\code{causalEffects_LL}}{ Lower limit of the \code{CI} of the (average) causal effects. It is calculated as the
(1-\code{CIboot})/2 \% quantile of the (average) causal effect across all explored DAGs }
\item{\code{causalEffects_UL}}{ Upper limit of the \code{CI} of the (average) causal effects. It is calculated as the
1-[(1-\code{CIboot})/2] \% quantile of the (average) causal effect across all explored DAGs }
\item{\code{group}}{ If \code{MrDAGcheck} in \code{MrDAG} algorithm contains the indices of the traits in \code{data} that
define the outcomes and the exposures, \code{group} coincides with this list }
\item{\code{ord}}{ Indices of the traits in \code{data} that specify the outcomes and the exposures. It might differ from
\code{group} if \code{ord} has been specified and it is different from \code{MrDAGcheck} }
\item{\code{BMA}}{ Logical option }
\item{\code{CI}}{ Level of the credible interval option } }
}
\description{
Estimation of the (average) causal effects under intervention on the exposures (targets) and measured on the outcomes
(responses) based on the Directed Acyclic Graphs (DAGs) explored by \code{MrDAG} algorithm
}
\examples{
# Example: Estimation of the (average) causal effects under intervention on lifestyle and 
# behavioural exposures, and measured on mental health phenotypes. 708 independent Instrumental 
# Variables (IVs) were selected to be associated at genome-wide significance with six lifestyle 
# and behavioural traits after pruning or clumping which are considered exposures of the risk 
# of seven mental health outcome phenotypes

# After loading the data set

data(LBT2MD_data)

# the indices of the traits that define the outcomes and exposures are provided in a list object

MrDAGcheck <- NULL
MrDAGcheck$Y_idx <- 1 : 7    # Mental health phenotypes
MrDAGcheck$X_idx <- 8 : 13   # Lifestyle and behavioural traits

# MrDAG algorithm is run to generate 1,000 posterior samples of all unknowns

output <- MrDAG(data = LBT2MD_data, 
                niter = 7500, burnin = 2500, thin = 5, tempMax = 20, pp = 0.01, 
                MrDAGcheck = MrDAGcheck, fileName = NULL)

# Finally, the posterior (average) causal effects and 90\% credible intervals are estimated

ord <- c(8 : 13, 1 : 7)
causalEffects <- get_causaleffects(output, ord = ord, CI = 0.90, progress = TRUE)
}
\references{
\insertAllCited{}
}
