% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_causaleffect.R
\name{get_causaleffect}
\alias{get_causaleffect}
\title{Estimation of the (average) marginal causal effects under intervention on an exposure}
\usage{
get_causaleffect(output, response, target, BMA = TRUE, CI = 0.95)
}
\arguments{
\item{output}{Output produced by \code{MrDAG} algorithm}

\item{response}{Trait (response) where the effect of the intervention on another trait (target) is measured. The index refers to
position of the trait (response) in the output produced by \code{MrDAG} algorithm}

\item{target}{Trait (target) under intervention. The index refers to position of the trait (response) in the output produced by
\code{MrDAG} algorithm}

\item{BMA}{If \code{TRUE}, Bayesian Model Averaging of the estimated (average) causal effect across all explored DAGs in the
visited Completed Partially DAG (CPDAG) (\insertCite{Chickering2002;textual}{MrDAG}) or Essential Graph (EG)
(\insertCite{Andersson1997;textual}{MrDAG}) is performed}

\item{CI}{Level (\code{0.95} default) of the credible interval (CI) of the (average) causal effect. It is calculated based on
suitable quantiles of the estimated (average) causal effect across all explored DAGs}
}
\value{
The value returned is a list object \code{list(causalEffect, causalEffect_LL, causalEffect_UL, group, BMA, CI)}
\itemize{
\item{\code{causalEffect}}{ Estimate of the (average) causal effect under intervention on an trait (target) on another
one (response) based on the DAGs explored by \code{MrDAG} algorithm. Feedback loop causal effect of the same trait is not
allowed (NA). Note that the estimated causal effect corresponds to the marginal causal effect of a target on a response
\emph{without} considering the effects of the others traits }
\item{\code{causalEffect_LL}}{ Lower limit of the \code{CI} of the (average) causal effect. It is calculated as the
(1-\code{CIboot})/2 \% quantile of the (average) causal effect across all explored DAGs }
\item{\code{causalEffect_UL}}{ Upper limit of the \code{CI} of the (average) causal effect. It is calculated as the
1-[(1-\code{CIboot})/2] \% quantile of the (average) causal effect across all explored DAGs }
\item{\code{group}}{ If \code{MrDAGcheck} in \code{MrDAG} algorithm contains the indices of the traits in \code{data} that
define the outcomes and the exposures, \code{group} coincides with this list }
\item{\code{BMA}}{ Logical option }
\item{\code{CI}}{ Level of the credible interval option } }
}
\description{
Estimation of the (average) marginal causal effect under intervention on a trait (target) and measured on another one
(response) based on the Directed Acyclic Graphs (DAGs) explored by \code{MrDAG} algorithm
}
\examples{
# Example: Estimation of the (average) causal effects under intervention on lifestyle and 
# behavioural exposures, and measured on mental health phenotypes. 708 independent Instrumental 
# Variables (IVs) were selected to be associated at genome-wide significance with six lifestyle 
# and behavioural traits after pruning or clumping which are considered exposures of the risk 
# of seven mental health outcome phenotypes

# After loading the data set

data(LBT2MD_data)

# the indices of the traits that define the outcomes and exposures are provided in a list object

MrDAGcheck <- NULL
MrDAGcheck$Y_idx <- 1 : 7    # Mental health phenotypes
MrDAGcheck$X_idx <- 8 : 13   # Lifestyle and behavioural traits

# MrDAG algorithm is run to generate 1,000 posterior samples of all unknowns

output <- MrDAG(data = LBT2MD_data, 
                niter = 7500, burnin = 2500, thin = 5, tempMax = 20, pp = 0.01, 
                MrDAGcheck = MrDAGcheck, fileName = NULL)

# Finally, the posterior (average) causal effects and credible intervals (CI) of the intervention
# on ALC and measured on SCZ are estimated

causalEffect <- get_causaleffect(output, 7, 8)   # Intervention on ALC(8) and measured on SCZ(7)

causalEffect <- get_causaleffect(output, 7, c(4, 8))   # Intervention on ALC(8) and measured on 
                                                       # SCZ(7) after conditioning on BD(4)
}
\references{
\insertAllCited{}
}
