% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MrDAG.R
\name{MrDAG}
\alias{MrDAG}
\title{MrDAG: Bayesian causal graphical model for joint Mendelian randomization analysis of multiple exposures and outcomes}
\usage{
MrDAG(
  data,
  niter,
  burnin,
  thin,
  pp = 0.05,
  a = NULL,
  U = NULL,
  tempMax = 10,
  MrDAGcheck = NULL,
  fileName = "MrDAG_object",
  filePath = NULL,
  saveMemory = FALSE,
  seed = 31122021
)
}
\arguments{
\item{data}{Number of observations (IVs in a summary-level MR design) times the number of traits (both outcomes and exposures).
There is a restriction in the order of appearance of the traits in the \code{data} matrix: The group of outcomes have to appear
first, followed by the group of exposures. See also \code{MrDAGcheck} argument}

\item{niter}{Number of MCMC iterations (including burn-in)}

\item{burnin}{Number of MCMC iterations to be discarded as burn-in}

\item{thin}{Parameter that defines how often the MCMC output should be stored, i.e., at every thin-th iteration}

\item{pp}{Prior probability of edge inclusion (\code{0.05} default) between each pair of nodes (vertices) in the graph}

\item{a}{Degrees of freedom of the Wishart prior distribution on the precision matrix, i.e., the inverse of the covariance matrix
between the traits (\code{m} default, where \code{m} is the total number of outcomes and exposures)}

\item{U}{(Proportional to the) expected value of the Wishart prior distribution on the precision matrix, i.e., the inverse of the
covariance matrix between the traits. Specifically, proportional to an m-dimensional diagonal matrix as default, where \code{m} is
the total number of outcomes and exposures}

\item{tempMax}{Annealing parameter \code{T} used to facilitate the convergence of the MCMC algorithm to the target distribution
(\code{10} default). Temperature \code{1/T} increases linearly during the burn-in until \code{T=1} at the end of the burn-in}

\item{MrDAGcheck}{List object that contains the indices of the traits in \code{data} that are defined as outcomes and exposures.
If \code{NULL} (default), no partition of the traits between outcomes and exposures is specified and the \code{MrDAG} algorithm
performs structure learning between all traits without constraints. Note that if so, this procedure does not correspond to an MR
analysis since valid instruments for the exposures are invalid for the outcomes and vice versa}

\item{fileName}{Name of the file for \code{MrDAG} output object (\code{"MrDAG_object"} default). If \code{NULL}, the ouput is not
saved}

\item{filePath}{Directory where \code{MrDAG} object is saved. If the directory does not exist, the \code{MrDAG} algorithm will
create it by using the current working directory as the root directory. If the directory is not specified (\code{NULL} default),
\code{MrDAG} object is saved in the current working directory}

\item{saveMemory}{If logical \code{FALSE} (default), the visited graph and the posterior draws from the modified Cholesky
decomposition (L,D) (\insertCite{Zuber2025;textual}{MrDAG} and \insertCite{Castelletti2021;textual}{MrDAG}) are stored as an array,
otherwise they are stored as a list object}

\item{seed}{Seed to be used in the initialisation of the MCMC algorithm (\code{3112021} default)}
}
\value{
The value returned is a list object \code{list(graphs, L,D, logMargLik, validPropMrDAG, acceptPropDAG, timeMrDAG, 
hyperPar, samplerPar, opt)}
\itemize{
\item{\code{graphs}}{ Explored DAGs that belong to the learned Completed Partially DAGs (CPDAGs)
(\insertCite{Chickering2002;textual}{MrDAG}) or Essential Graphs (EGs) (\insertCite{Andersson1997;textual}{MrDAG}). The
class of \code{graphs} depends on the \code{saveMemory} option. The number of explored DAGs corresponds to the number of
thinned (\code{thin}) MCMC iterations (excluding burn-in) }
\item{\code{L}}{ Posterior samples of the lower triangular matrix of the modified Cholesky decomposition (L,D)
(\insertCite{Zuber2025;textual}{MrDAG} and \insertCite{Castelletti2021;textual}{MrDAG}) }
\item{\code{D}}{ Posterior samples of the diagonal matrix of the modified Cholesky decomposition (L,D)
(\insertCite{Zuber2025;textual}{MrDAG} and \insertCite{Castelletti2021;textual}{MrDAG}) }
\item{\code{logMargLik}}{ Log-marginal likelihood of explored DAGs which belong to the Markov Equivalent Classes whose unique
representative chain graphs are the EGs learned during MCMC iterations (including burn-in) without thinning }
\item{\code{validPropMrDAG}}{ If \code{MrDAGcheck} is different from \code{NULL}, the proportion of proposed DAGs that comply
with the partial ordering (\insertCite{Perkovic2017;textual}{MrDAG} implied by the partition of the traits between
exposures and outcomes with directed causal effects only from the former to the latter }
\item{\code{acceptPropDAG}}{ Proportion of proposed DAGs that are accepted by the Metropolis-Hastings ratio after
burn-in. If \code{MrDAGcheck} is different from \code{NULL}, \code{acceptPropDAG} corresponds to the proposed DAGs that
comply with the partial ordering}
\item{\code{timeMrDAG}}{ Time in minutes employed by the \code{MrDAG} algorithm to analyse the data }
\item{\code{hyperPar}}{ List of the hyper-parameters \code{list(w, a, U, tempMax)} and, if specified, \code{MrDAGcheck} list }
\item{\code{samplerPar}}{ List of parameters used in the MCMC algorithm \code{list(niter, burnin, thin)} }
\item{\code{opt}}{ List of options used in \code{MrDAG} algorithm \code{list(saveMemory, seed)} } }
}
\description{
Markov chain Monte Carlo (MCMC) implementation of Bayesian multivariable, multi-response Mendelian randomization (MR)
model for summary-level data with structure learning and causal effects estimation. The directionality of the causal effects
between the exposures and the outcomes is assumed known, i.e., the exposures can only be potential causes of the outcomes and no
reverse causation is allowed
}
\details{
For details regarding the model and the algorithm, see \insertCite{Zuber2025;textual}{MrDAG}
}
\examples{
# Example: Analysis of lifestyle and behavioural exposures that might impact mental health 
# phenotypes. 708 independent Instrumental Variables (IVs) were selected to be associated at 
# genome-wide significance with six lifestyle and behavioural traits after pruning or clumping,
# which are considered exposures of the risk of seven mental health outcome phenotypes

# After loading the data set

data(LBT2MD_data)

# the indices of the traits that define the outcomes and exposures are provided in a list object

MrDAGcheck <- NULL
MrDAGcheck$Y_idx <- 1 : 7    # Mental health phenotypes
MrDAGcheck$X_idx <- 8 : 13   # Lifestyle and behavioural traits

# MrDAG algorithm is run to generate 1,000 posterior samples of all unknowns with memory savings

output <- MrDAG(data = LBT2MD_data, 
                niter = 7500, burnin = 2500, thin = 5, tempMax = 20, pp = 0.01, 
                MrDAGcheck = MrDAGcheck, fileName = NULL)


}
\references{
\insertAllCited{}
}
